Move clr overlays to sub-modules (#1094)
