+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Remember to bring your dependencies up to date with `./scripts/venvinstall.sh` w
 - Minor: Added user-specific cooldown to playsound module (#888, #1006)
 - Minor: Add \$(broadcaster) variable (#925, #1076)
 - Minor: Duel winrate now takes `duels_won` into consideration if winrate is equal (#1079)
+- Minor: Moved CLR-based modules to a sub-folder. (#1094)
 - Bugfix: Paid timeouts will now only timeout once. (#993)
 - Bugfix: Handle new format of P&SL lists (#988, #989)
 - Bugfix: Fixed league rank module not working at all. (#990)
