+++ pajbot/modules/__init__.py
@@ -25,14 +25,16 @@ from pajbot.modules.chat_alerts.newchatteralert import NewChatterAlertModule
 from pajbot.modules.chat_alerts.subalert import SubAlertModule
 from pajbot.modules.chat_alerts.raidalert import RaidAlertModule
 from pajbot.modules.chatters_refresh import ChattersRefreshModule
+from pajbot.modules.clr_overlay import CLROverlayModule
+from pajbot.modules.clr_overlay.emotecombo import EmoteComboModule
+from pajbot.modules.clr_overlay.emotesonscreen import EmotesOnScreenModule
+from pajbot.modules.clr_overlay.showemote import ShowEmoteModule
 from pajbot.modules.deck import DeckModule
 from pajbot.modules.duel import DuelModule
 from pajbot.modules.dummy import DummyModule
 from pajbot.modules.eightball import EightBallModule
 from pajbot.modules.emote_timeout import EmoteTimeoutModule
 from pajbot.modules.emote_limit import EmoteLimitModule
-from pajbot.modules.emotecombo import EmoteComboModule
-from pajbot.modules.emotesonscreen import EmotesOnScreenModule
 from pajbot.modules.givepoints import GivePointsModule
 from pajbot.modules.hsbet import HSBetModule
 from pajbot.modules.lastfm import LastfmModule
@@ -68,7 +70,6 @@ from pajbot.modules.quests.winraffle import WinRaffleQuestModule
 from pajbot.modules.raffle import RaffleModule
 from pajbot.modules.repspam import RepspamModule
 from pajbot.modules.roulette import RouletteModule
-from pajbot.modules.showemote import ShowEmoteModule
 from pajbot.modules.slotmachine import SlotMachineModule
 from pajbot.modules.subscriber_fetch import SubscriberFetchModule
 from pajbot.modules.top import TopModule
@@ -90,6 +91,7 @@ available_modules = [
     ChattersRefreshModule,
     CheckModModule,
     ClipCommandModule,
+    CLROverlayModule,
     DBManageModule,
     DebugModule,
     DeckModule,
