+++ pajbot/modules/clr_overlay/emotecombo.py
@@ -3,15 +3,17 @@ import logging
 from pajbot.managers.handler import HandlerManager
 from pajbot.modules import BaseModule
 from pajbot.modules.base import ModuleSetting
+from pajbot.modules.clr_overlay import CLROverlayModule
 
 log = logging.getLogger(__name__)
 
 
 class EmoteComboModule(BaseModule):
     ID = __name__.split(".")[-1]
-    NAME = "Emote Combos - Overlay"
+    NAME = "Emote Combos"
     DESCRIPTION = "Shows emote combos on the CLR pajbot overlay"
     CATEGORY = "Feature"
+    PARENT_MODULE = CLROverlayModule
     SETTINGS = [
         ModuleSetting(
             key="min_emote_combo",
