+++ pajbot/modules/clr_overlay/emotesonscreen.py
@@ -4,15 +4,17 @@ import random
 from pajbot.managers.handler import HandlerManager
 from pajbot.modules import BaseModule
 from pajbot.modules import ModuleSetting
+from pajbot.modules.clr_overlay import CLROverlayModule
 
 log = logging.getLogger(__name__)
 
 
 class EmotesOnScreenModule(BaseModule):
     ID = __name__.split(".")[-1]
-    NAME = "Emotes on Screen - CLR"
+    NAME = "Emotes on Screen"
     DESCRIPTION = "Shows one or more emotes on screen per message"
     CATEGORY = "Feature"
+    PARENT_MODULE = CLROverlayModule
     SETTINGS = [
         ModuleSetting(
             key="emote_whitelist",
